requirements.txt
python-docx>=0.8.11
pdfplumber>=0.10.0
tqdm>=4.65.0
